Add exclude option

+++ clean-branch.py
@@ -5,18 +5,20 @@ import subprocess
 import re
 
 
-def clean(repo_path, deleted_remote=False, local_only=False, dry_run=True):
+def clean(repo_path, deleted_remote=False, local_only=False, excluded_folders = '', dry_run=True):
     os.chdir(repo_path)
 
     # Call external command 'git branch -vv' and get the command output
     cmd = 'git branch -vv'
     output = subprocess.check_output(cmd, shell=True).decode('utf-8')
-    print(output)
+    excluded_folder_list = [] if excluded_folders is None else excluded_folders.split(';')
     # Split the output by line
     lines = output.split('\n')
     # Loop through each line
     branch_list = []
     for line in lines:
+        if line.startswith('*'):
+            continue
         if line.strip() == '':
             continue
         branch_pattern = re.compile(r'[*+\s]\s'
@@ -33,25 +35,30 @@ def clean(repo_path, deleted_remote=False, local_only=False, dry_run=True):
         #                              branch_match.group('remote_branch'),
         #                              branch_match.group('remote_status') == 'gone'))
         need_delete = False
+        local_path = branch_match.group('local_path')
+        if local_path is not None:
+            local_folder = os.path.basename(local_path)
+            if local_folder in excluded_folder_list:
+                continue
         if deleted_remote and branch_match.group('remote_status') == 'gone':
             need_delete = True
         if not need_delete and local_only and branch_match.group('remote_branch') is None:
             need_delete = True
         if need_delete:
             branch_list.append(branch_match.group('local_branch'))
-            local_path = branch_match.group('local_path')
             if local_path is not None:
                 if dry_run:
                     print('Would delete local path: {}'.format(local_path))
                 else:
                     print('Deleting......: {}'.format(local_path))
                     shutil.rmtree(local_path)
-    if dry_run:
-        print('Would delete branches: {}'.format(' '.join(branch_list)))
-    else:
-        print('Deleting......: {}'.format(' '.join(branch_list)))
-        subprocess.check_output('git worktree prune', shell=True)
-        subprocess.check_output('git branch -D {}'.format(' '.join(branch_list)), shell=True)
+    if len(branch_list) > 0:
+        if dry_run:
+            print('Would delete branches: {}'.format(' '.join(branch_list)))
+        else:
+            print('Deleting......: {}'.format(' '.join(branch_list)))
+            subprocess.check_output('git worktree prune', shell=True)
+            subprocess.check_output('git branch -D {}'.format(' '.join(branch_list)), shell=True)
 
 
 if __name__ == '__main__':
@@ -64,8 +71,10 @@ if __name__ == '__main__':
                         help='clean branches only exists locally', default=False)
     parser.add_argument('-d', '--dry-run', action='store_true',
                         help='clean branches only exists locally', default=False)
+    parser.add_argument('-e', '--exclude', action='store',
+                        help='folder names combined with ; which will not be excluded from clean')
     args = parser.parse_args()
     if args.repo_path is None:
         print('Repo path not provided.')
         exit(1)
-    clean(args.repo_path, args.gone_remote, args.local_only, args.dry_run)
+    clean(args.repo_path, args.gone_remote, args.local_only, args.exclude, args.dry_run)
